chore(nav): rename /cars route to /vehicles (#150)
Rename member-facing fleet page, API routes, and hooks:
- app/cars/ → app/vehicles/
- app/api/cars/ → app/api/vehicles/
- hooks/use-cars.ts → hooks/use-vehicles.ts
- hooks/use-car-state.ts → hooks/use-vehicle-state.ts
- Update nav-drawer, proxy ADMIN_ONLY_PAGES, prewarm, e2e helpers
