+++ app/api/vehicles/[id]/last-state/route.ts
