+++ app/api/vehicles/[id]/route.ts
