+++ app/api/vehicles/route.ts
