+++ app/fuel/fuel-form.tsx
@@ -9,8 +9,8 @@ import { ReceiptUpload } from "@/components/receipt-upload";
 import { LocationPicker } from "@/components/location-picker";
 import { calcPricePerLiter } from "@/lib/formulas";
 import { usePeople } from "@/hooks/use-people";
-import { useCars } from "@/hooks/use-cars";
-import { useLastCarState } from "@/hooks/use-car-state";
+import { useCars } from "@/hooks/use-vehicles";
+import { useLastCarState } from "@/hooks/use-vehicle-state";
 import { useMe } from "@/hooks/use-me";
 import { useOnlineState } from "@/lib/offline/online-state";
 import type { FuelFillup, FuelFillupInput } from "@/types";
