+++ hooks/use-vehicle-state.ts
@@ -6,6 +6,6 @@ export function useLastCarState(carId: number | undefined) {
   return useQuery<CarState | null>({
     queryKey: ["car-state", carId],
     enabled: Boolean(carId),
-    queryFn: () => apiFetch<CarState | null>(`/api/cars/${carId}/last-state`),
+    queryFn: () => apiFetch<CarState | null>(`/api/vehicles/${carId}/last-state`),
   });
 }
